Digging items
Just added the digging items with their correct chance of being dug related to each map. So the player will only find items that are supposed to be in that map.

This chance of being dug is calculated based on an item's potential of being found, which I chose from 1 to 5. So, depending on the items that are in the map and how high is their potential of being find, the game automatically calculates the chance of finding each item, so that they sum up to 100%. This way, the player always digs exactly 1 item and the items that have higher potential with respect to the other items in the map appear more frequently.

This all happens while the game is being loaded, so it doesn't affect the performance in playtime.