Spells implemented and tested
Uff, this was a marathon to get through! Implemented all spells that needed implementation and tested them all.
Fixed a couple bugs on the way and optimized other things.
Also fixed some stuff in the .csv and .json files.
Resolved a bunch of TODOs.
Onward to resolve the rest and finish this!