Live being status
Generalized status so now every attribute can have their own and made them all last according to time (not frame) \o/
On top of that, made blood and poison be applied at a determined time rate (again, not every frame), which makes a huge difference!
With these changes, the code is shorter, more general, and does more!
This is something I've been wanting to do for the longest time, and now I finally had the tools (TimeCounter class) and strategy to do it.
Now, one of the most complex and messiest parts of the game is organized, generalized and working fine.