Fullscreen mode!
Made it possible to resize the window to other resolutions, even to fullscreen.
This requested big refactoring, even separating the JFrame class from the main class (which I had coming for a while).
More refactoring in on the way! I just wanted to do the minimum necessary for adding this feature. Now, onto make it better!